wCCD: Improve error handling (#232)
Display error messages in the application instead of logging them to console or showing alert popups. This adds a bunch of error fields to the various components' state.

Error messages got a pass and the connection logic was migrated from the app component into the `WithWalletConnection` library component.

This improves the UX and solves the following issues:

- Browser Wallet:

  - Clean up event handlers when "disconnecting" (i.e. discarding) the connector. This avoids interference between reconnects.
  - Show error message when browser wallet isn't available.

- WalletConnect:

  - Disconnect is now detected properly.
  - Ensure that we no longer get stuck in "waiting for user" state when the modal is closed without connecting.

- RPC errors are improved and displayed in the UI. This makes it as clear as it can be if the connected account belongs to the wrong chain.

The UI now only shows the "transaction status" and "admin instance" fields when the is a connection as there otherwise isn't anything useful to show in them. As part of this the "lean how to make a wrapper like this" link was moved down next to the version.