eSealing: Fix label and error resetting (#249)
The label around the file input covered a lot of elements below the component, such that e.g. clicking on the computed file hash opened the file dialog.

Also, once a file hash computation returned an error, this error would never be reset even if it was successful afterwards (can trigger error on Ubuntu by selecting a text file, editing the file, then computing the hash). Fixed by resetting the error on success.